Small Caps Workaround Fix
Small Caps contexts now use a manually-configured set of labels that permit fine-tuning of the ratio between uppercase and lowercase, instead of the built-in smallcaps which is built on the fixed HTML specification.
Refactored the horrid HTMLFONT section to use a macro (function) instead of the garbage that was there before.  Created macros.nde to contain this and other macros.