Replacement fixes
Altered tower and king replacement with digits (i.e. 3T or 4K) to be case-sensitive; this permits "4th" etc to be used without issue.
Added automatic en dash insertion when a - is in front of a digit.  Added manual en dash replacement of -- and em dash replacement of ---.
Added support for a custom newline character in addition to <br>.