Title/Subtitle Auto-Accent Replacements and other fixes
Switched auto-accent replacements to use an explicit REPLACE instead of HTMLKEY, which is opaque and performed with timing that we have no control over.  All card text replacements now use this new version, including title, subtitle, card type, subtype, game text, and lore.  Replacements are defined as a pair of Text_Find and Text_Replace sequences in a format that should be a little easier for less-technical users to grok.
Configurable newline character modified to explicitly disallow various special characters within nanDECK, such as [, ], and |.
Twilight symbol replacements modified to wrap the replacement in <typesymbol> tags, which gets italicized helper text to work with them.
Altered the image size filter to Lanczos for hopefully an easy fidelity upgrade.
Altered position and size of collector's info to be more compatible with longer coll infos.
Altered site game text to not cut off a bold 'damage +1' on the bottom line.